synthtask-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
requests==2.31.0
python-multipart==0.0.6
google-generativeai==0.3.2

# MongoDB
motor==3.3.2
pymongo==4.6.1

# PostgreSQL
databases==0.9.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# Security
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# SQLAlchemy sync driver for metadata.create_all
psycopg2-binary==2.9.11